requirements.txt
absl-py==2.3.0
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
actuator-msgs==0.0.1
altair==5.5.0
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
bond==4.1.2
builtin-interfaces==1.2.1
cachetools==4.2.2
cartographer-ros-msgs==2.0.9002
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
composition-interfaces==1.2.1
control-msgs==4.8.0
controller-manager==2.50.0
controller-manager-msgs==2.50.0
cv-bridge==3.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.8.0
diagnostic-updater==4.0.4
domain-coordinator==0.10.0
dwb-msgs==1.1.18
dynamixel-sdk==3.8.3
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.3.0
flatbuffers==25.2.10
gast==0.6.0
gazebo-msgs==3.9.0
generate-parameter-library-py==0.4.0
geometry-msgs==4.8.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gps-msgs==2.0.4
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
image-geometry==3.2.1
interactive-markers==2.3.2
Jinja2==3.1.6
joblib==1.5.1
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
laser-geometry==2.4.0
launch==1.0.8
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
libclang==18.1.1
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
message-filters==4.3.7
ml_dtypes==0.5.1
mocap4r2-control-msgs==0.0.7
namex==0.1.0
narwhals==1.42.0
nav-2d-msgs==1.1.18
nav-msgs==4.8.0
nav2-common==1.1.18
nav2-msgs==1.1.18
nav2-simple-commander==1.0.0
nltk==3.9.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
osrf-pycommon==2.1.6
packaging==24.2
pandas==2.3.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-qt-binding==1.1.2
python-telegram-bot==13.7
pytz==2025.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
resource-retriever==3.1.3
rich==14.0.0
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.17
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2controlcli==2.50.0
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpds-py==0.25.1
rpyutils==0.2.1
rqt==1.1.7
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-controller-manager==2.50.0
rqt-dotgraph==0.0.4
rqt-gauges==0.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-joint-trajectory-controller==2.45.0
rqt-moveit==1.0.1
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.6
rqt-robot-steering==1.0.1
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-tf-tree==1.0.5
rqt-topic==1.5.0
rsa==4.9.1
scripts==3.9.0
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
six==1.17.0
slam-toolbox==2.6.10
smclib==4.1.2
smmap==5.0.2
sniffio==1.3.1
sros2==0.10.6
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
streamlit==1.45.1
teleop-twist-keyboard==2.4.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
theora-image-transport==2.5.3
toml==0.10.2
topic-monitor==0.20.5
tornado==6.1
tqdm==4.67.1
trajectory-msgs==4.8.0
turtlebot3-autorace-camera==1.2.2
turtlebot3-autorace-detect==1.2.2
turtlebot3-autorace-mission==1.2.2
turtlebot3-example==2.2.9
turtlebot3-msgs==2.4.0
turtlebot3-teleop==2.2.9
turtlesim==1.4.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
unique-identifier-msgs==2.2.1
uritemplate==4.2.0
urllib3==2.4.0
vision-msgs==4.1.1
visualization-msgs==4.8.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
xacro==2.0.13
